No clue what I'm doing

+++ tool/travis.sh
@@ -15,7 +15,7 @@ pushd $PKG
 EXIT_CODE=0
 
 # TODO remove this
-echo $1
+echo "WTF: $1"
 
 #escapedPath="$(echo $PWD | sed 's/\//\\\//g')"
 escapedPath="$(echo $1 | sed 's/\//\\\//g')"
@@ -39,13 +39,14 @@ while (( "$#" )); do
     pub upgrade || exit $?
     echo -e '\033[1mTASK: test\033[22m'
     echo -e 'pub run test'
-    pub run test || EXIT_CODE=$?
-
-    pub global run coverage:collect_coverage --port=8111 --out=coverage.json --wait-paused --resume-isolates
+    nohup pub global run coverage:collect_coverage --port=8111 -o coverage.json --resume-isolates --wait-paused &
+    dart --pause-isolates-on-exit --enable-vm-service=8111 "test/all_tests.dart" || EXIT_CODE=$?
     pub global run coverage:format_coverage --packages=.packages -i coverage.json --report-on lib --lcov --out lcov.info
-
-    sed "s/^SF:.*lib/SF:$escapedPath\/lib/g" lcov.info >> $2/lcov.info
-    rm lcov.info
+    if [ -f "lcov.info" ]; then
+      # combine line coverage info from package tests to a common file
+      sed "s/^SF:.*lib/SF:$escapedPath\/lib/g" lcov.info >> $2/lcov.info
+      rm lcov.info
+    fi
     rm -f coverage.json
     ;;
   flutter_analyze) echo
@@ -58,8 +59,11 @@ while (( "$#" )); do
     echo -e '\033[1mTASK: flutter test\033[22m'
     echo -e 'flutter test'
     flutter test --coverage || EXIT_CODE=$?
-#    sed "s/^SF:lib/SF:$escapedPath\/lib/g" coverage/lcov.info >> $2/lcov.info
-#    rm -rf "coverage"
+    if [ -d "coverage" ]; then
+      # combine line coverage info from package tests to a common file
+      sed "s/^SF:lib/SF:$escapedPath\/lib/g" coverage/lcov.info >> $2/lcov.info
+      rm -rf "coverage"
+    fi
     ;;
   *) echo -e "\033[31mNot expecting TASK '${TASK}'. Error!\033[0m"
     EXIT_CODE=1
